Remove VehicleUnit
