+++ examples/supply_chain/common/balance_calculator.py
@@ -242,14 +242,6 @@ class BalanceSheetCalculator:
 
         return storage_product_step_cost
 
-    def _calc_vehicle(self, tick: int) -> np.ndarray:
-        # loss = cost * payload
-        vehicle_step_cost = -1 * (
-            self._get_attributes("vehicle", "payload", tick)
-            * self._get_attributes("vehicle", "unit_transport_cost", tick)
-        )
-        return vehicle_step_cost
-
     def _calc_product_distribution(self, tick: int) -> Tuple[np.ndarray, np.ndarray]:
         # product distribution profit = check order * price
         product_distribution_step_profit = (
