+++ maro/simulator/scenarios/supply_chain/__init__.py
@@ -3,19 +3,17 @@
 
 from .actions import ConsumerAction, ManufactureAction, SupplyChainAction
 from .datamodels import (
-    ConsumerDataModel, DistributionDataModel, ManufactureDataModel, SellerDataModel, StorageDataModel, VehicleDataModel
+    ConsumerDataModel, DistributionDataModel, ManufactureDataModel, SellerDataModel, StorageDataModel
 )
 from .facilities import FacilityBase, RetailerFacility, SupplierFacility, WarehouseFacility
 from .units import (
-    ConsumerUnit, DistributionUnit, ExtendUnitBase, ManufactureUnit, ProductUnit, SellerUnit, StorageUnit, UnitBase,
-    VehicleUnit
+    ConsumerUnit, DistributionUnit, ExtendUnitBase, ManufactureUnit, ProductUnit, SellerUnit, StorageUnit, UnitBase
 )
 
 __all__ = [
     "ConsumerAction", "ManufactureAction", "SupplyChainAction",
     "ConsumerDataModel", "DistributionDataModel", "ManufactureDataModel", "SellerDataModel", "StorageDataModel",
-    "VehicleDataModel",
     "FacilityBase", "RetailerFacility", "SupplierFacility", "WarehouseFacility",
     "ConsumerUnit", "DistributionUnit", "ExtendUnitBase", "ManufactureUnit", "ProductUnit", "SellerUnit",
-    "StorageUnit", "UnitBase", "VehicleUnit",
+    "StorageUnit", "UnitBase",
 ]
