+++ maro/simulator/scenarios/supply_chain/units/__init__.py
@@ -10,4 +10,3 @@ from .product import ProductUnit, ProductUnitInfo, StoreProductUnit
 from .seller import OuterSellerUnit, SellerDemandMixin, SellerUnit, SellerUnitInfo
 from .storage import StorageUnit, StorageUnitInfo
 from .unitbase import BaseUnitInfo, UnitBase
-from .vehicle import VehicleUnit
