+++ maro/simulator/scenarios/supply_chain/world.py
@@ -275,24 +275,6 @@ class World:
 
         return unit_instance
 
-    def _build_distribution_unit(self, facility: FacilityBase, config: dict) -> DistributionUnit:
-        unit_def: EntityDef = self.configs.entity_defs[config["class"]]
-        assert issubclass(unit_def.class_type, DistributionUnit)
-
-        distribution_unit = self._build_unit(facility, facility, config)
-        assert isinstance(distribution_unit, DistributionUnit)
-        distribution_unit.children = []
-        distribution_unit.vehicles = defaultdict(list)
-
-        vehicle_configs: Dict[str, dict] = config.get("vehicles", {})
-        for vehicle_type, vehicle_config in vehicle_configs.items():
-            for _ in range(vehicle_config.get("number", 1)):
-                vehicle_unit = self._build_unit(facility, distribution_unit, vehicle_config["config"])
-                distribution_unit.children.append(vehicle_unit)
-                distribution_unit.vehicles[vehicle_type].append(vehicle_unit)
-
-        return distribution_unit
-
     def _build_product_units(self, facility: FacilityBase, config: dict) -> Dict[int, ProductUnit]:
         """Generate product unit by sku information.
 
@@ -403,10 +385,7 @@ class World:
 
             # Build children Units.
             for child_name, child_conf in facility_conf["children"].items():
-                if child_name == "distribution":
-                    child = self._build_distribution_unit(facility, child_conf)
-                else:
-                    child = self._build_unit(facility=facility, parent=facility, config=child_conf)
+                child = self._build_unit(facility=facility, parent=facility, config=child_conf)
                 setattr(facility, child_name, child)
 
             # Build ProductUnits.
