+++ tests/data/supply_chain/case_01/config.yml
@@ -44,12 +44,11 @@ medium_storage: &medium_storage
 
 
 train_config: &train
-  class: "VehicleUnit"
   config:
     patient: 3
 
 normal_distribution: &normal_distribution
-  vehicles:
+  config:
     train:
       number: 2
       config: *train
