+++ tests/data/supply_chain/case_02/config.yml
@@ -63,13 +63,12 @@ multiple_storage: &multiple_storage
 
 
 train_config: &train
-  class: "VehicleUnit"
   config:
     patient: 3
     unit_transport_cost: 12
 
 normal_distribution: &normal_distribution
-  vehicles:
+  config:
     train:
       number: 2
       config: *train
